Move AddWebApi call to regular IDnnStartup
This ensures that it's called at the right time relative to all other DI
configuration.
